requirements.txt
pdfplumber>=0.11
google-api-python-client>=2.100
google-auth-oauthlib>=1.0
google-auth-httplib2>=0.2
python-dotenv>=1.0
pytest>=8.0
